# Conflicts:
#	ms-core/src/main/java/com/hq/ecmp/mscore/service/IOrderInfoService.java
#	ms-core/src/main/java/com/hq/ecmp/mscore/service/impl/OrderInfoServiceImpl.java

+++ ms-core/src/main/java/com/hq/ecmp/mscore/service/impl/EnterpriseCarTypeInfoServiceImpl.java
@@ -1,14 +1,24 @@
 package com.hq.ecmp.mscore.service.impl;
 
+import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
+
+import com.google.common.collect.Lists;
 import com.hq.common.utils.DateUtils;
+import com.hq.ecmp.constant.CarConstant;
+import com.hq.ecmp.mscore.domain.CarInfo;
 import com.hq.ecmp.mscore.domain.EcmpEnterpriseInfo;
 import com.hq.ecmp.mscore.domain.EcmpUser;
 import com.hq.ecmp.mscore.domain.EnterpriseCarTypeInfo;
+import com.hq.ecmp.mscore.dto.CarTypeDTO;
+import com.hq.ecmp.mscore.mapper.CarInfoMapper;
 import com.hq.ecmp.mscore.mapper.EcmpEnterpriseInfoMapper;
 import com.hq.ecmp.mscore.mapper.EcmpUserMapper;
 import com.hq.ecmp.mscore.mapper.EnterpriseCarTypeInfoMapper;
 import com.hq.ecmp.mscore.service.IEnterpriseCarTypeInfoService;
+import com.hq.ecmp.mscore.vo.CarTypeVO;
+import com.hq.ecmp.mscore.vo.PageResult;
 import org.apache.commons.collections.CollectionUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -28,6 +38,8 @@ public class EnterpriseCarTypeInfoServiceImpl implements IEnterpriseCarTypeInfoS
     private EcmpUserMapper ecmpUserMapper;
     @Autowired
     private EcmpEnterpriseInfoMapper ecmpEnterpriseInfoMapper;
+    @Autowired
+    private CarInfoMapper carInfoMapper;
 
     /**
      * 查询【请填写功能名称】
@@ -98,9 +110,18 @@ public class EnterpriseCarTypeInfoServiceImpl implements IEnterpriseCarTypeInfoS
      * @return 结果
      */
     @Override
-    public int deleteEnterpriseCarTypeInfoById(Long carTypeId)
-    {
-        return enterpriseCarTypeInfoMapper.deleteEnterpriseCarTypeInfoById(carTypeId);
+    public int deleteEnterpriseCarTypeInfoById(Long carTypeId) throws Exception {
+        //判断该车型下是否绑定车辆 如有绑定则无法删除
+        CarInfo carInfo = CarInfo.builder().carTypeId(carTypeId).build();
+        List<CarInfo> carInfos = carInfoMapper.selectCarInfoList(carInfo);
+        if(carInfos.size() != 0){
+            throw new Exception("请先移除类型下绑定的车辆");
+        }
+        int i = enterpriseCarTypeInfoMapper.deleteEnterpriseCarTypeInfoById(carTypeId);
+        if(i != 1){
+            throw new Exception("删除失败");
+        }
+        return i;
     }
 
     /**
@@ -128,5 +149,94 @@ public class EnterpriseCarTypeInfoServiceImpl implements IEnterpriseCarTypeInfoS
         //查询公司车型
         List<EnterpriseCarTypeInfo> enterpriseCarTypeInfos = enterpriseCarTypeInfoMapper.selectEnterpriseCarTypeInfoList(enterpriseCarTypeInfo);
         return enterpriseCarTypeInfos;
+    }
+
+    /**
+     * 新增车型
+     * @param carDto
+     * @param userId
+     */
+    @Override
+    public void saveCarType(CarTypeDTO carDto, Long userId) throws Exception {
+        EnterpriseCarTypeInfo enterpriseCarTypeInfo = new EnterpriseCarTypeInfo();
+        enterpriseCarTypeInfo.setEnterpriseId(carDto.getEnterpriseId());
+        enterpriseCarTypeInfo.setName(carDto.getName());
+        enterpriseCarTypeInfo.setLevel(carDto.getLevel());
+        // 初始化状态为生效  状态   S000   生效中    S444   失效中
+        enterpriseCarTypeInfo.setStatus(CarConstant.START_CAR_TYPE);
+        enterpriseCarTypeInfo.setCreateBy(String.valueOf(userId));
+        enterpriseCarTypeInfo.setCreateTime(new Date());
+        int i = enterpriseCarTypeInfoMapper.insertEnterpriseCarTypeInfo(enterpriseCarTypeInfo);
+        if(i != 1){
+            throw new Exception("新增车型失败");
+        }
+    }
+
+    /**
+     * 修改车型
+     * @param carDto
+     * @param userId
+     */
+    @Override
+    public void updateCarType(CarTypeDTO carDto, Long userId) throws Exception {
+        EnterpriseCarTypeInfo enterpriseCarTypeInfo = new EnterpriseCarTypeInfo();
+        enterpriseCarTypeInfo.setCarTypeId(carDto.getCarTypeId());
+        enterpriseCarTypeInfo.setName(carDto.getName());
+        enterpriseCarTypeInfo.setLevel(carDto.getLevel());
+        enterpriseCarTypeInfo.setUpdateBy(String.valueOf(userId));
+        enterpriseCarTypeInfo.setUpdateTime(new Date());
+        int i = enterpriseCarTypeInfoMapper.updateEnterpriseCarTypeInfo(enterpriseCarTypeInfo);
+        if(i != 1){
+            throw new Exception();
+        }
+    }
+
+    /**
+     * 查询企业车型列表
+     * @param enterpriseId
+     * @return
+     */
+    @Override
+    public List<CarTypeVO> getCarTypeList(Long enterpriseId) {
+        EnterpriseCarTypeInfo enterpriseCarTypeInfo = new EnterpriseCarTypeInfo();
+        enterpriseCarTypeInfo.setEnterpriseId(enterpriseId);
+        List<EnterpriseCarTypeInfo> enterpriseCarTypeInfos = enterpriseCarTypeInfoMapper.selectEnterpriseCarTypeInfoList(enterpriseCarTypeInfo);
+        CarTypeVO carTypeVO = null;
+        List<CarTypeVO> list = Lists.newArrayList();
+        for (EnterpriseCarTypeInfo carTypeInfo : enterpriseCarTypeInfos) {
+            carTypeVO = CarTypeVO.builder().carTypeId(carTypeInfo.getCarTypeId())
+                    .countryCarTypeId(carTypeInfo.getCountryCarTypeId())
+                    .level(carTypeInfo.getLevel())
+                    .name(carTypeInfo.getLevel())
+                    .enterpriseId(carTypeInfo.getEnterpriseId())
+                    .build();
+            list.add(carTypeVO);
+        }
+        return list;
+    }
+
+    /**
+     * 车型排序（交换位置）
+     * @param mainCarTypeId
+     * @param targetCarTypeId
+     */
+    @Override
+    public void sortCarType(Long mainCarTypeId, Long targetCarTypeId,Long userId) throws Exception {
+        EnterpriseCarTypeInfo mainCarTypeInfo = enterpriseCarTypeInfoMapper.selectEnterpriseCarTypeInfoById(mainCarTypeId);
+        EnterpriseCarTypeInfo targetCarTypeInfo = enterpriseCarTypeInfoMapper.selectEnterpriseCarTypeInfoById(targetCarTypeId);
+        mainCarTypeInfo.setLevel(targetCarTypeInfo.getLevel());
+        mainCarTypeInfo.setUpdateTime(new Date());
+        mainCarTypeInfo.setUpdateBy(String.valueOf(userId));
+        targetCarTypeInfo.setLevel(mainCarTypeInfo.getLevel());
+        targetCarTypeInfo.setUpdateBy(String.valueOf(userId));
+        targetCarTypeInfo.setUpdateTime(new Date());
+        int i = enterpriseCarTypeInfoMapper.updateEnterpriseCarTypeInfo(mainCarTypeInfo);
+        if(i != 1){
+            throw new Exception();
+        }
+        int j = enterpriseCarTypeInfoMapper.updateEnterpriseCarTypeInfo(targetCarTypeInfo);
+        if(j != 1){
+            throw new Exception();
+        }
     }
 }
